+++ eval-config.nix
@@ -52,6 +52,7 @@ let
       system.nssModules = dummy;
       system.requiredKernelConfig = dummy;
       system.stateVersion = optionValue (if legacyInstallDirs then "21.11" else "22.05");
+      systemd.oomd = dummy;
       ids.gids.keys = dummy;
       ids.uids.systemd-coredump = dummy;
       ids.gids.systemd-journal = dummy;